Move commandParser and commandResult to cli
