+++ src/cli/commandResult.js
