+++ src/commands/cloneExam.js
@@ -5,7 +5,7 @@ var mkdirp        = require("mkdirp");
 // var request       = require("request");
 var moment        = require("moment");
 var Promise       = require("bluebird");
-var CommandResult = require("../commandResult");
+var CommandResult = require("../cli/commandResult");
 
 var CloneChallengeCommand = require("./cloneChallenge");
